refactor(rp): split equipment.rs into per-device-type equipment/ module
Mechanical move with no behavior change. Split the 1,531-line
services/rp/src/equipment.rs into the per-device-type subtree
prescribed by docs/services/rp.md §"Module Structure":

  equipment/
    mod.rs               EquipmentRegistry + impl (new, status,
                         find_*, validate_site), the Status types,
                         SITE_MATCH_TOLERANCE_DEG, and re-exports
                         of the per-device Entry types so existing
                         crate::equipment::CameraEntry imports keep
                         working unchanged.
    alpaca.rs            Generic Alpaca client glue:
                         build_alpaca_client, retry_connect_attempt,
                         AttemptOutcome enum, CONNECT_ATTEMPTS /
                         GET_DEVICES_TIMEOUT consts. Tests for both
                         helpers live here.
    camera.rs            CameraEntry + connect_camera.
    filter_wheel.rs      FilterWheelEntry + connect_filter_wheel.
    cover_calibrator.rs  CoverCalibratorEntry + connect_cover_calibrator.
    focuser.rs           FocuserEntry + connect_focuser + tests
                         (failure modes, success path, end-to-end
                         registry test, ok_focuser_router fixture).
    mount.rs             MountEntry + connect_mount + tests
                         (failure modes, success, end-to-end registry,
                         the eight validate_site_* tests, all the
                         mount-router stub fixtures including
                         mount_router_with_site / _without_site).
    test_support.rs      AlpacaStub + spawn_stub — the only stub
                         helper genuinely shared between focuser
                         and mount tests. Gated on #[cfg(test)] and
                         only visible inside the crate.

The skip prescribed by the design doc — equipment/safety_monitor.rs —
is left out: there is no SafetyMonitor wrapper code yet (config
keeps a typed-loose `safety_monitors: Vec<Value>`). Add the file
when the wrapper lands.

Submodule visibility: connect_* functions and alpaca.rs's helpers are
pub(super) so EquipmentRegistry::new can call them across the module
boundary. Entry struct fields stay pub (callers like routes.rs's
test fixtures construct registries directly).

Validation: cargo build passes, all 386 rp library tests still pass,
cargo fmt clean. Test-path namespacing is the only visible diff
(equipment::tests::connect_focuser_X → equipment::focuser::tests::connect_focuser_X).

Second in a 3-step refactor sequence: config.rs (4aa1939),
equipment.rs (this), then mcp.rs (per-tool-category split using
rmcp 1.4 multi-router pattern).
